Add unit test for SmtpServer
Add a single unit test, which send a single static email to an
instance of the SmtpServer and compares the recieved email with the
sent email.
Implement From trait for SmtpEmail to simplify the setup for the test.
Adjust the function SmtpServer::new() to bind to an address given as
an argument, to allow other ports than 25 for testing purposes.