Improve error handling
Remove all calls to Result.expect(), which correspond to errors we can
recover from, and replace them with better handling.

To allow for a better handling of errors which occour inside the
smtp_server::MailHandler we have to remove the channel for sending the
received email to the context of the handle_mail_comm() function.
Instead the received email is stored inside the MailHandler and only a
reference to this Handler is given to the Session, which allows us to
retreive the email after closing the Session.

Refactor the Error enum by removing unnecessary variants.
Implement Display trait for Error enum.